+++ src/app/(admin)/admin/skilltree/page.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import SkillTreeManager from '../../../../components/SkillTreeManager';
+
+export default function ManageSkillTreePage() {
+  return <SkillTreeManager />;
+}